PROTON-27 : Provides a more robust environment for developing the Ruby bindings.
Fixes the Rspec target name to be "rspec" rather than "spec".
Added a TODO and ChangeLog file for the Ruby language bindings.
Updated the Rakefile to include these in the gem.
Added new files to the gitignore file for the Ruby bindings.
Added a rake target to generate the Ruby documentation.
Added support for using SimpleCov to provide test coverage.

git-svn-id: https://svn.apache.org/repos/asf/qpid/proton/trunk@1394227 13f79535-47bb-0310-9956-ffa450edef68