PROTON-445: Perl install now honors CMAKE_INSTALL_PREFIX
The language bindings will now honor this prefix, both the default
value and a user-specified value.

Developers can override it, though, by specifying PERL_INSTALL_PREFIX
at the command line as well. This will allow for installing the Perl
bindings to a different directory if so desired.

Also updated the installation pieces in CMake to simplify it.

git-svn-id: https://svn.apache.org/repos/asf/qpid/proton/trunk@1547578 13f79535-47bb-0310-9956-ffa450edef68