PROTON-445: Dynamic languages honor CMAKE_INSTALL_PREFIX
All languages are installed to $CMAKE_INSTALL_PREFIX/bindings/$LANG by
default.

If the ASK_ALL macro is set to 1 at the command line, then each
language is interrogated as to the location of where they will be
installed, and that path modified wit the install prefix.

Individual languages can be told to interrogate for the install path
with:

  ASK_[LANG]=1

git-svn-id: https://svn.apache.org/repos/asf/qpid/proton/trunk@1571719 13f79535-47bb-0310-9956-ffa450edef68